fix(control): accept Foxy float32 sport telemetry
Allow numbers.Real values when projecting the fixed three-axis SportModeState velocity array. The deployed Unitree Foxy message materializes those elements as numpy.float32 rather than Python float.

Keep booleans, non-finite values, oversized axes, malformed array lengths, and non-numeric values rejected. Add an exact numpy.float32 regression case based on the stationary hardware observation.

+++ tests/test_go2_bridge.py
@@ -4,6 +4,8 @@ import unittest
 from pathlib import Path
 from types import SimpleNamespace
 
+import numpy as np
+
 from robot_dashboard.control import ACTION_GUARD_S, SAFE_ACTIONS
 from robot_dashboard.go2_bridge import (
     API_MOVE,
@@ -194,7 +196,8 @@ class Go2BridgeCoreTests(unittest.TestCase):
             SimpleNamespace(
                 mode=5,
                 gait_type=3,
-                velocity=[0.1051234, -0.02, 0.0],
+                # Foxy materializes fixed float arrays as numpy.float32 values.
+                velocity=[np.float32(0.1051234), np.float32(-0.02), np.float32(0.0)],
                 error_code=SPORT_MODE_STATE_MAX_ERROR_CODE,
             ),
             now=10.0,
